api/requirements.txt
alembic==1.8.1
mysqlclient==2.0.1
opencv-contrib-python==4.2.0.34
opencv-python==4.2.0.34
Pillow==7.2.0
PyJWT==2.3.0
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
requests==2.26.0
SQLAlchemy==1.4.44
ulid-py==0.0.7
uvicorn==0.20.0
fastapi==0.88.0
itsdangerous==2.1.2
gunicorn==20.1.0
jinja2==3.1.2
line-bot-sdk==2.3.0